requirements.txt
streamlit
easyocr
pytesseract
pandas
Pillow
matplotlib
